еще генерация данных + организация по файлам
