+++ src/mock/gen-point-data.js
@@ -1,47 +1,7 @@
-import {getRandomPositiveNumber} from '../utils/random-gen.js';
+import dayjs from 'dayjs';
+import {POINT_TYPES, CITIES, OFFER_NAMES, PRICES} from './data-sources.js';
+import {getRandomPositiveNumber, getRandomElement} from '../utils/random-gen.js';
 
-const DESTINATIONS = [
-  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-  'Cras aliquet varius magna, non porta ligula feugiat eget.',
-  'Fusce tristique felis at fermentum pharetra.',
-  'Aliquam id orci ut lectus varius viverra.',
-  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
-  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
-  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
-  'Sed sed nisi sed augue convallis suscipit in sed felis.',
-  'Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus.',
-  'In rutrum ac purus sit amet tempus.',
-];
-
-const POINT_TYPES = ['Taxi', 'Bus', 'Train', 'Ship', 'Drive', 'Flight', 'Check-in', 'Sightseeing', 'Restaurant'];
-
-const CITIES = ['Amsterdam', 'Chamonix','Geneva','New-York','Berlin','London','Paris','Moscow','City-17'];
-
-const OFFER_NAMES = ['Add luggage', 'Switch to comfort class', 'Add meal', 'Choose seats', 'Travel by train'];
-
-const PRICES = ['30', '100', '15', '5', '40']; //четко соответсвует по порядку именам офферов
-
-const getRandomElement = (array) => array[getRandomPositiveNumber(0, array.length-1)];
-
-//destionations
-const findDestinationSentence = (usedDestinations) => {
-  let sentence = getRandomElement(DESTINATIONS);
-  while (usedDestinations.includes(sentence)) { //проверка на уникальность
-    sentence = getRandomElement(DESTINATIONS);
-  }
-  usedDestinations.push(sentence);
-  return sentence;
-};
-
-const generateDestinationsText = () => {
-  const sentencesAmount = getRandomPositiveNumber(1, 5);
-  const usedDestinations = [];
-  let destinationText = '';
-  for (let i = 0; i < sentencesAmount; i++) {
-    destinationText += `${findDestinationSentence(usedDestinations)} `;
-  }
-  return destinationText;
-};
 
 //point-type
 const pointType = getRandomElement(POINT_TYPES);
@@ -49,9 +9,8 @@ const pointType = getRandomElement(POINT_TYPES);
 //cities
 const city = getRandomElement(CITIES);
 
-//offers - пока непонятно что здесь
-
-const createOffer = () => ({ //походу не работает такое создание объекта
+//offers
+const createOffer = () => ({
   _name: getRandomElement(OFFER_NAMES),
   currency: '&plus;&euro;&nbsp',
   get price() {
@@ -74,6 +33,17 @@ const generateOffers = () => {
   }
   return usedOffers;
 };
-const gen = generateOffers();
-console.log(generateOffers());
 
+//date
+const generateDate = () => {
+  const dateDifference = getRandomPositiveNumber(1, 15);
+  return dayjs().add(dateDifference, 'day').toDate();
+};
+
+//favorite
+const isFavorite = Boolean(getRandomPositiveNumber(0, 1));
+
+//price
+const generatePrice = () => Math.ceil(getRandomPositiveNumber(2, 80))*10;
+
+//time
